requirements.txt
flask==0.12.1
redis==2.10.5
PyYAML==3.12
